Drop SqlDialectTrait.
Move the methods in the Driver class itself.
